Move default IPC path for dsrouter component into dotnet-dsrouter.
